+++ actionview/lib/action_view/helpers/tags/time_zone_select.rb
@@ -1,9 +1,13 @@
 # frozen_string_literal: true
 
+require "action_view/helpers/tags/select_renderer"
+
 module ActionView
   module Helpers
     module Tags # :nodoc:
       class TimeZoneSelect < Base # :nodoc:
+        include SelectRenderer
+
         def initialize(object_name, method_name, template_object, priority_zones, options, html_options)
           @priority_zones = priority_zones
           @html_options   = html_options
