+++ actionview/lib/action_view/helpers/tags/weekday_select.rb
@@ -1,9 +1,13 @@
 # frozen_string_literal: true
 
+require "action_view/helpers/tags/select_renderer"
+
 module ActionView
   module Helpers
     module Tags # :nodoc:
       class WeekdaySelect < Base # :nodoc:
+        include SelectRenderer
+
         def initialize(object_name, method_name, template_object, options, html_options)
           @html_options = html_options
